GQL-12 Page Layout Fixes (#6)
* GQL-12 feat(layout): add schema

* GQL-12 feat(layout): add tabs

* GQL-12 fix(layout): change schema btn to div

* GQL-12 feat(layout): add play btn

* GQL-12 fix(layout): fix console error

* GQL-12 fix(develop): fix header position
