+++ src/index.scss
@@ -21,10 +21,10 @@ main {
   justify-content: center;
   align-items: center;
   width: 100%;
-  height: 100vh;
 }
 
 .cm-theme-light,
 .cm-theme-dark {
   width: 100%;
+  flex-grow: 1;
 }
